refactor(workflow-history): Rewire external consumers to v2 shared deps
Point workflow-page, formatters, and summary tests at workflow-history-v2
for fixtures, filter constants, pending-event types, and the history tab
entry wrapper so v1 is no longer a cross-app dependency.
