+++ src/utils/data-formatters/format-pending-workflow-history-event/__tests__/index.test.ts
@@ -7,7 +7,7 @@ import {
   pendingActivityTaskStartEventWithStartedState,
   pendingDecisionTaskStartEvent,
   pendingDecisionTaskStartEventWithStartedState,
-} from '@/views/workflow-history/__fixtures__/workflow-history-pending-events';
+} from '@/views/workflow-history-v2/__fixtures__/workflow-history-pending-events';
 
 import formatPendingWorkflowHistoryEvent from '..';
