+++ src/utils/data-formatters/format-pending-workflow-history-event/index.ts
@@ -1,5 +1,5 @@
 import logger from '@/utils/logger';
-import { type PendingHistoryEvent } from '@/views/workflow-history/workflow-history.types';
+import { type PendingHistoryEvent } from '@/views/workflow-history-v2/workflow-history-v2.types';
 
 import {
   getFormatPendingEventSchema,
