+++ src/views/workflow-page/config/workflow-page-query-params.config.ts
@@ -2,10 +2,10 @@ import {
   type PageQueryParam,
   type PageQueryParamMultiValue,
 } from '@/hooks/use-page-query-params/use-page-query-params.types';
-import { HISTORY_EVENT_FILTER_STATUSES } from '@/views/workflow-history/workflow-history-filters-status/workflow-history-filters-status.constants';
-import { type HistoryEventFilterStatus } from '@/views/workflow-history/workflow-history-filters-status/workflow-history-filters-status.types';
-import { WORKFLOW_HISTORY_EVENT_FILTERING_TYPES } from '@/views/workflow-history/workflow-history-filters-type/workflow-history-filters-type.constants';
-import { type WorkflowHistoryEventFilteringType } from '@/views/workflow-history/workflow-history-filters-type/workflow-history-filters-type.types';
+import { HISTORY_EVENT_FILTER_STATUSES } from '@/views/workflow-history-v2/workflow-history-filters-status/workflow-history-filters-status.constants';
+import { type HistoryEventFilterStatus } from '@/views/workflow-history-v2/workflow-history-filters-status/workflow-history-filters-status.types';
+import { WORKFLOW_HISTORY_EVENT_FILTERING_TYPES } from '@/views/workflow-history-v2/workflow-history-filters-type/workflow-history-filters-type.constants';
+import { type WorkflowHistoryEventFilteringType } from '@/views/workflow-history-v2/workflow-history-filters-type/workflow-history-filters-type.types';
 
 const workflowPageQueryParamsConfig: [
   PageQueryParamMultiValue<
